separated dashbaord components in the matters page
